feat: 原color组件 setup 并 改名为colorSelector
